# specs/001-ratemyprof-india-platform/contracts/api-specification.yaml
openapi: 3.0.3
info:
  title: RateMyProf India API
  description: API for rating and discovering professors in Indian colleges
  version: 0.1.0
  contact:
    name: RateMyProf India
    url: https://github.com/NihaallX/ratemyprof.in
servers:
  - url: https://api.ratemyprof.in/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
paths:
...
  /professors:
    get:
      summary: Search professors by name, college, or department
      operationId: searchProfessors
      tags: [professors]
      parameters:
        - name: q
          in: query
          description: Search query (professor name)
          schema:
            type: string
            minLength: 2
            maxLength: 100
        - name: college_id
          in: query
          description: Filter by specific college
          schema:
            type: string
            format: uuid
        - name: college_name
          in: query
          description: Filter by college name (fuzzy match)
          schema:
            type: string
            maxLength: 200
        - name: department
          in: query
          description: Filter by department
          schema:
            type: string
            maxLength: 100
        - name: state
          in: query
          description: Filter by Indian state
          schema:
            type: string
            maxLength: 50
        - name: limit
          in: query
          description: Maximum results to return
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Professor search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  professors:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProfessorSummary'
                  total:
                    type: integer
                    description: Total matching professors
                  has_more:
                    type: boolean
                    description: Whether more results available
        '400':
          $ref: '#/components/responses/BadRequest'
  /professors/{professor_id}:
    get:
      summary: Get professor profile with reviews
      operationId: getProfessor
      tags: [professors]
      parameters:
        - name: professor_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: reviews_limit
          in: query
          description: Number of reviews to include
          schema:
            type: integer
            minimum: 0
            maximum: 100
            default: 20
        - name: reviews_offset
          in: query
          description: Reviews pagination offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Professor profile with reviews
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfessorProfile'
        '404':
          $ref: '#/components/responses/NotFound'

  # Reviews
...
  /auth/signup:
    post:
      summary: Register new user account
      operationId: signup
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignup'
      responses:
        '201':
          description: User account created, verification email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Account created. Please check your email for verification."
                  user_id:
                    type: string
                    format: uuid
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: Login with email and password
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/verify:
    post:
      summary: Verify email address with OTP
      operationId: verifyEmail
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                otp_code:
                  type: string
                  pattern: '^[0-9]{6}$'
              required: [email, otp_code]
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email verified successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid or expired verification code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Colleges
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    College:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        city:
          type: string
          maxLength: 100
        state:
          type: string
          maxLength: 50
        college_type:
          type: string
          enum: [university, college, institute]
        established_year:
          type: integer
          minimum: 1800
          maximum: 2025
          nullable: true
      required: [id, name, city, state, college_type]
    ProfessorSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
          maxLength: 50
        last_name:
          type: string
          maxLength: 50
        known_as:
          type: string
          maxLength: 50
          nullable: true
        department:
          type: string
          maxLength: 100
        college:
          $ref: '#/components/schemas/College'
        avg_ratings:
          $ref: '#/components/schemas/AverageRatings'
        total_reviews:
          type: integer
          minimum: 0
      required: [id, first_name, last_name, department, college, avg_ratings, total_reviews]
    ProfessorProfile:
      allOf:
        - $ref: '#/components/schemas/ProfessorSummary'
        - type: object
          properties:
            profile_photo_url:
              type: string
              format: uri
              nullable: true
            subjects_taught:
              type: array
              items:
                type: string
              nullable: true
            reviews:
              type: array
              items:
                $ref: '#/components/schemas/Review'
            reviews_total:
              type: integer
            reviews_has_more:
              type: boolean
    AverageRatings:
      type: object
      properties:
        clarity:
          type: number
          format: float
          minimum: 1.0
          maximum: 5.0
          nullable: true
        helpfulness:
          type: number
          format: float
          minimum: 1.0
          maximum: 5.0
          nullable: true
        workload:
          type: number
          format: float
          minimum: 1.0
          maximum: 5.0
          nullable: true
        engagement:
          type: number
          format: float
          minimum: 1.0
          maximum: 5.0
          nullable: true
        overall:
          type: number
          format: float
          minimum: 1.0
          maximum: 5.0
          nullable: true
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        professor_id:
          type: string
          format: uuid
        author_name:
          type: string
          maxLength: 100
          description: "Display name or 'Anonymous'"
        ratings:
          type: object
          properties:
            clarity:
              type: integer
              minimum: 1
              maximum: 5
            helpfulness:
              type: integer
              minimum: 1
              maximum: 5
            workload:
              type: integer
              minimum: 1
              maximum: 5
            engagement:
              type: integer
              minimum: 1
              maximum: 5
          required: [clarity, helpfulness, workload, engagement]
        review_text:
          type: string
          maxLength: 2000
          nullable: true
        semester_taken:
          type: string
          maxLength: 50
          nullable: true
        course_taken:
          type: string
          maxLength: 100
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_flagged:
          type: boolean
        flags_count:
          type: integer
          minimum: 0
      required: [id, professor_id, author_name, ratings, created_at, updated_at, is_flagged, flags_count]

    # Request Models
...
    UserSignup:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 8
          maxLength: 128
        name:
          type: string
          maxLength: 100
          nullable: true
        college_id:
          type: string
          format: uuid
          nullable: true
      required: [email, password]
    UserLogin:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: "JWT access token"
        token_type:
          type: string
          example: "Bearer"
        expires_in:
          type: integer
          description: "Token expiration in seconds"
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            name:
              type: string
              nullable: true
            is_verified:
              type: boolean
            is_moderator:
              type: boolean
          required: [id, email, is_verified, is_moderator]
      required: [access_token, token_type, expires_in, user]

    # Moderation Models
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: "Error type"
        message:
          type: string
          description: "Human-readable error message"
        details:
          type: object
          nullable: true
          description: "Additional error context"
      required: [error, message]
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "validation_error"
            message: "Invalid input data"
            details:
              field_errors:
                email: ["Invalid email format"]
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Professor not found"
...